Rename OpentracingSpanManager to SpanManager.
This fixes #34.
