+++ opentracing-span-propagation/src/main/java/nl/talsmasoftware/context/opentracing/OpentracingSpanManager.java
@@ -15,52 +15,9 @@
  */
 package nl.talsmasoftware.context.opentracing;
 
-import io.opentracing.Scope;
-import io.opentracing.ScopeManager;
-import io.opentracing.Span;
-import io.opentracing.util.GlobalTracer;
-import nl.talsmasoftware.context.Context;
-import nl.talsmasoftware.context.ContextManager;
-
-import java.util.concurrent.atomic.AtomicBoolean;
-
 /**
  * @author Sjoerd Talsma
+ * @deprecated Renamed to {@code SpanManager}
  */
-public class OpentracingSpanManager implements ContextManager<Span> {
-
-    @Override
-    public Context<Span> getActiveContext() {
-        ScopeManager scopeManager = GlobalTracer.get().scopeManager();
-        return new ScopeContext(scopeManager.active(), true);
-    }
-
-    @Override
-    public Context<Span> initializeNewContext(final Span span) {
-        ScopeManager scopeManager = GlobalTracer.get().scopeManager();
-        return new ScopeContext(span == null ? null : scopeManager.activate(span, false), false);
-    }
-
-    private static class ScopeContext implements Context<Span> {
-        private final Scope scope;
-        private final AtomicBoolean closed;
-
-        private ScopeContext(Scope scope, boolean alreadyClosed) {
-            this.scope = scope;
-            this.closed = new AtomicBoolean(alreadyClosed);
-        }
-
-        @Override
-        public Span getValue() {
-            return scope == null ? null : scope.span();
-        }
-
-        @Override
-        public void close() {
-            if (closed.compareAndSet(false, true) && scope != null) {
-                scope.close();
-            }
-        }
-    }
-
+public class OpentracingSpanManager extends SpanManager {
 }
