+++ opentracing-span-propagation/src/main/resources/META-INF/services/nl.talsmasoftware.context.ContextManager
@@ -1,1 +1,1 @@
-nl.talsmasoftware.context.opentracing.OpentracingSpanManager
+nl.talsmasoftware.context.opentracing.SpanManager
