+++ pom.xml
@@ -16,7 +16,9 @@
     limitations under the License.
 
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
 
@@ -196,12 +198,12 @@
                             </docletArtifact>
                             <docencoding>${project.reporting.outputEncoding}</docencoding>
                             <useStandardDocletOptions>true</useStandardDocletOptions>
-                            <additionalparam>
-                                -umlImageFormat SVG
-                                -umlImageDirectory images
-                                -umlCommand "hide empty fields"
-                                -umlCommand "hide empty methods"
-                            </additionalparam>
+                            <additionalOptions>
+                                <additionalOption>-umlImageFormat SVG</additionalOption>
+                                <additionalOption>-umlImageDirectory images</additionalOption>
+                                <additionalOption>-umlCommand "hide empty fields"</additionalOption>
+                                <additionalOption>-umlCommand "hide empty methods"</additionalOption>
+                            </additionalOptions>
                         </configuration>
                     </execution>
                 </executions>
@@ -212,7 +214,9 @@
                 <version>${license-maven-plugin.version}</version>
                 <executions>
                     <execution>
-                        <goals><goal>check</goal></goals>
+                        <goals>
+                            <goal>check</goal>
+                        </goals>
                         <phase>compile</phase>
                     </execution>
                 </executions>
